+++ modules/TGraphDB/src/main/java/edu/buaa/act/gephi/plugin/task/TimeDependentDijkstraOneTransactionAsyncTask.java
@@ -51,7 +51,7 @@ public class TimeDependentDijkstraOneTransactionAsyncTask extends Traverse{
     private Map<Long,org.gephi.graph.api.Node> tgraphNode2GephiNode = new HashMap<Long,org.gephi.graph.api.Node>();
     private List<Long> path;
     private List<Integer> timeList;
-    private int pathRealLength = 0;
+    private double pathRealLength = 0;
     private long searchCount=0;
 
 
@@ -173,7 +173,7 @@ public class TimeDependentDijkstraOneTransactionAsyncTask extends Traverse{
                 if(edge!=null){
                     edge.setColor(pathColor);
                     edge.setWeight(2f);
-                    pathRealLength += (Integer) edge.getAttribute("road_length");
+                    pathRealLength += (Double) edge.getAttribute("road_length");
                     System.out.println("edge has direct.");
                 }else{
                     System.out.println("no edge between nodes! "+parent+"--"+child);
@@ -181,7 +181,7 @@ public class TimeDependentDijkstraOneTransactionAsyncTask extends Traverse{
             }else{
                 edge.setColor(pathColor);
                 edge.setWeight(2f);
-                pathRealLength += (Integer) edge.getAttribute("road_length");
+                pathRealLength += (Double) edge.getAttribute("road_length");
             }
             child = parent;
         }
@@ -194,7 +194,7 @@ public class TimeDependentDijkstraOneTransactionAsyncTask extends Traverse{
         Edge edge = model.getGraph().getEdge(begin, end);
         edge.setColor(pathColor);
         edge.setWeight(2f);
-        pathRealLength += (Integer) edge.getAttribute("road_length");
+        pathRealLength += (Double) edge.getAttribute("road_length");
         begin.setSize(3f);
         begin.setColor(Color.RED);
         //reverse array
